requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3